Combined ConsumerConfig with RabbitConfig and deleted ConsumerConfig. Removed RabbitTemplate bean creation being as that is for sending rabbit messages. Changed ConsumerHandler to MessageService.
