+++ index/src/main/groovy/ncei/catalog/config/RabbitConfig.groovy
@@ -1,19 +1,24 @@
-package ncei.catalog.amqp
+package ncei.catalog.config
 
+import ncei.catalog.service.MessageService
 import org.springframework.amqp.core.AmqpAdmin
+import org.springframework.amqp.core.Queue
 import org.springframework.amqp.rabbit.connection.CachingConnectionFactory
 import org.springframework.amqp.rabbit.connection.ConnectionFactory
 import org.springframework.amqp.rabbit.core.RabbitAdmin
+import org.springframework.amqp.rabbit.core.RabbitTemplate
+import org.springframework.amqp.rabbit.listener.SimpleMessageListenerContainer
+import org.springframework.amqp.rabbit.listener.adapter.MessageListenerAdapter
 import org.springframework.amqp.support.converter.DefaultClassMapper
 import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter
 import org.springframework.amqp.support.converter.MessageConverter
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 
 @Configuration
 class RabbitConfig {
-
   @Value('${rabbitmq.connectionfactory.virtualHost}')
   String virtualHost
 
@@ -26,6 +31,20 @@ class RabbitConfig {
   @Value('${rabbitmq.connectionfactory.password}')
   String password
 
+  final static String queueName = 'index-consumer'
+
+  @Autowired
+  private MessageService scrapingResultHandler
+
+  @Bean
+  Queue listenerQueue() {
+    new Queue(queueName, true, false, true)
+  }
+  @Bean
+  AmqpAdmin amqpAdmin() {
+    new RabbitAdmin(connectionFactory())
+  }
+
   @Bean
   ConnectionFactory connectionFactory() {
     CachingConnectionFactory connectionFactory = new CachingConnectionFactory(host)
@@ -36,8 +55,17 @@ class RabbitConfig {
   }
 
   @Bean
-  AmqpAdmin amqpAdmin() {
-    new RabbitAdmin(connectionFactory())
+  SimpleMessageListenerContainer listenerContainer() {
+    SimpleMessageListenerContainer container = new SimpleMessageListenerContainer()
+    container.setConnectionFactory(connectionFactory())
+    container.setQueueNames(queueName)
+    container.setMessageListener(messageListenerAdapter())
+    container
+  }
+
+  @Bean
+  MessageListenerAdapter messageListenerAdapter() {
+    new MessageListenerAdapter(scrapingResultHandler, jsonMessageConverter())
   }
 
   @Bean
